+++ BE/src/stock/enum/market-type.ts
