+++ BE/src/stock/index/dto/stock-index-list-chart.element.dto.ts
