+++ BE/src/stock/index/dto/stock-index-response-element.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { StockIndexValueElementDto } from './stock.index.value.element.dto';
-import { StockIndexListChartElementDto } from './stock.index.list.chart.element.dto';
+import { StockIndexValueElementDto } from './stock-index-value-element.dto';
+import { StockIndexListChartElementDto } from './stock-index-list-chart.element.dto';
 
 export class StockIndexResponseElementDto {
   @ApiProperty({ description: '실시간 값', type: StockIndexValueElementDto })
