+++ BE/src/stock/index/dto/stock-index-value-element.dto.ts
