+++ BE/src/stock/index/interface/stock-index.interface.ts
