+++ BE/src/stock/index/stock-index.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { StockIndexController } from './stock.index.controller';
-import { StockIndexService } from './stock.index.service';
-import { KoreaInvestmentModule } from '../../koreaInvestment/korea.investment.module';
+import { StockIndexController } from './stock-index.controller';
+import { StockIndexService } from './stock-index.service';
+import { KoreaInvestmentModule } from '../../koreaInvestment/korea-investment.module';
 import { SocketModule } from '../../websocket/socket.module';
 
 @Module({
