+++ BE/src/stock/topfive/interface/stock-topfive.interface.ts
