+++ BE/src/stock/topfive/stock-topfive.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { StockTopfiveController } from './stock.topfive.controller';
-import { StockTopfiveService } from './stock.topfive.service';
-import { KoreaInvestmentModule } from '../../koreaInvestment/korea.investment.module';
+import { StockTopfiveController } from './stock-topfive.controller';
+import { StockTopfiveService } from './stock-topfive.service';
+import { KoreaInvestmentModule } from '../../koreaInvestment/korea-investment.module';
 
 @Module({
   imports: [ConfigModule, KoreaInvestmentModule],
